refactor. 폴더명 수정 및 폴더 구조 동기화
