+++ src/entites/comment/model/types.ts
