+++ src/entites/post/api.ts
@@ -1,4 +1,4 @@
-import { NewPost, PaginatedPosts, Post } from "./types"
+import { NewPost, PaginatedPosts, Post } from "./model/types"
 
 export const fetchPosts = async ({ limit, skip }: { limit: number; skip: number }): Promise<PaginatedPosts> => {
   const response = await fetch(`/api/posts?limit=${limit}&skip=${skip}`)
