+++ src/entites/user/lib/getUserFromId.ts
@@ -1,4 +1,4 @@
-import { User } from "../types"
+import { User } from "../model/types"
 
 export const getUserFromId = (users: User[], id: number) => {
   return users.find((user) => user.id === id)
