+++ src/entites/user/model/types.ts
